Gestion de la production par seconde.

+++ app/src/main/java/com/example/makemoremeat/models/Game.kt
@@ -6,23 +6,23 @@ import kotlin.reflect.KProperty
 
 class Game : PropertyChangeAware() {
 
-    private val observer = {
-            property: KProperty<*>,
-            oldValue: Any,
-            newValue: Any -> propertyChangeSupport.firePropertyChange(property.name, oldValue, newValue)
+    private val observer = { property: KProperty<*>,
+                             oldValue: Any,
+                             newValue: Any ->
+        propertyChangeSupport.firePropertyChange(property.name, oldValue, newValue)
     }
 
-    var money: Long by Delegates.observable(0, observer)
+    var money: Double by Delegates.observable(0.0, observer)
     var moneyPerSecond = 0
     var fastUP = 1 // 1 = x1, 2 = x10, 3 = x25, 4 = xMax
-    var butchers:Array<Butcher> = arrayOf()
+    var butchers: Array<Butcher> = arrayOf()
 
     fun addButcher(butcher: Butcher) {
         butchers.plus(butcher)
     }
 
     fun hardReset() {
-        money = 0
+        money = 0.0
     }
 
 }
